chore: separate cloud only types from being exported generally
