+++ src/types/index.ts
@@ -7,7 +7,6 @@ export * from './auth'
 export * from './autoRefresh'
 export * from './billing'
 export * from './buckets'
-export * from './cloud'
 export * from './codemirror'
 export * from './colors'
 export * from './dashboards'
